+++ tools/npm-discards.js
@@ -7,7 +7,7 @@ var hasOwn = Object.prototype.hasOwnProperty;
 // This class encapsulates a structured specification of files and
 // directories that should be stripped from the node_modules directories
 // of Meteor packages during `meteor build`, as requested by calling
-// `Npm.discard` in package.js files.
+// `Npm.discard` in meteor-package.js files.
 function NpmDiscards() {
   assert.ok(this instanceof NpmDiscards);
   this.discards = {};